Add sqlalchemy and pessimistic database connections
MySQL has a habit of dropping connections after a period of time (usually
around 8 hours or so). In addition, sometimes the database server needs to get
restarted for whatever reason. This patch adds sqlalchemy with connection
pooling and recylcing support. In addition, a cursor wrapper function will
automatically create a new connection if it thinks the connection was dropped.