The sutta search and dictionary searches now more-or-less work. There is significant work to be done in regards to presentation, and some tweaking required. However the basic functionality is quite good.
The fulltext search is also making an appearance, but is not functional
in this submit - it basically works, but isn't hooked up.

Note that before the dictionary search works, it is nessecary to manually
run build_dict_db.py - that will build the required database. Give the
static nature of dictionaries I think it is adequate to require a manual rebild.