Add remaining titles for suttas and translations.
Note: there are some problems with suttas that are grouped together.  For
example /an1.1-10/pi shows the title of the first sutta as opposed to the
collection of suttas. This looks like to be a problem with how groupings of
suttas are represented in the database. Some other suttas like /dhp1-20/pi seem
to be showing the titles correctly.