fix bottom margin on dictionary result pages
add margin-bottom of 48px for .entries and .related_terms. For some reason there doesn't seem to be a container that we can apply margin-bottom to for the page as a whole. It would be better to have such a container and apply margin to that, but oh well. The margin-bottom for .entries is removed on small screens, see media-responsive.scss